COMP:Fixed warning about comparison between signed and unsigned integer expressions.

+++ VTK/Common/vtkVariantArray.cxx
@@ -834,7 +834,8 @@ void vtkVariantArray::DataElementChanged(vtkIdType id)
         return;
         }
 
-      if (this->Lookup->CachedUpdates.size() > this->GetNumberOfTuples()/10)
+      if (this->Lookup->CachedUpdates.size() >
+          static_cast<size_t>(this->GetNumberOfTuples()/10))
         {
         // At this point, just rebuild the full table.
         this->Lookup->Rebuild = true;
